Reorganize onboarding composables
