+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/composables/onboarding/CheckOnboarding.kt
@@ -1,4 +1,4 @@
-package app.jjerrell.meta.course.sample.littlelemon.composables
+package app.jjerrell.meta.course.sample.littlelemon.composables.onboarding
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
