+++ app/src/main/java/app/jjerrell/meta/course/sample/littlelemon/composables/onboarding/OnboardingPage.kt
@@ -1,4 +1,4 @@
-package app.jjerrell.meta.course.sample.littlelemon.composables
+package app.jjerrell.meta.course.sample.littlelemon.composables.onboarding
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Arrangement
